+++ packages/affine/data-view/src/property-presets/multi-select/define.ts
@@ -2,9 +2,9 @@ import { nanoid } from '@blocksuite/store';
 
 import type { SelectPropertyData } from '../select/define.js';
 
+import { getTagColor } from '../../core/component/tags/colors.js';
 import { type SelectTag, t } from '../../core/index.js';
 import { propertyType } from '../../core/property/property-config.js';
-import { getTagColor } from '../../core/utils/tags/colors.js';
 
 export const multiSelectPropertyType = propertyType('multi-select');
 export const multiSelectPropertyModelConfig =