+++ packages/blocks/src/attachment-block/index.ts
@@ -1,2 +1,8 @@
 export * from './attachment-block.js';
 export * from './attachment-service.js';
+export { attachmentViewToggleMenu } from './components/options.js';
+export {
+  type AttachmentEmbedConfig,
+  AttachmentEmbedConfigIdentifier,
+  AttachmentEmbedProvider,
+} from './embed.js';