+++ packages/blocks/src/root-block/edgeless/gfx-tool/pan-tool.ts
@@ -1,6 +1,7 @@
 import type { PointerEventState } from '@blocksuite/block-std';
 
-import { BaseTool } from '@blocksuite/block-std/gfx';
+import { on } from '@blocksuite/affine-shared/utils';
+import { BaseTool, MouseButton } from '@blocksuite/block-std/gfx';
 import { Signal } from '@preact/signals-core';
 
 export type PanToolOption = {
@@ -13,6 +14,10 @@ export class PanTool extends BaseTool<PanToolOption> {
   private _lastPoint: [number, number] | null = null;
 
   readonly panning$ = new Signal<boolean>(false);
+
+  override get allowDragWithRightButton(): boolean {
+    return true;
+  }
 
   override dragEnd(_: PointerEventState): void {
     this._lastPoint = null;
@@ -37,6 +42,36 @@ export class PanTool extends BaseTool<PanToolOption> {
   override dragStart(e: PointerEventState): void {
     this._lastPoint = [e.x, e.y];
     this.panning$.value = true;
+  }
+
+  override mounted(): void {
+    this.addHook('pointerDown', evt => {
+      const shouldPanWithMiddle = evt.raw.button === MouseButton.MIDDLE;
+
+      if (!shouldPanWithMiddle) {
+        return;
+      }
+
+      evt.raw.preventDefault();
+
+      const currentTool = this.controller.currentToolOption$.peek();
+      const restoreToPrevious = () => {
+        this.controller.setTool(currentTool);
+      };
+
+      this.controller.setTool('pan', {
+        panning: true,
+      });
+
+      const dispose = on(document, 'pointerup', evt => {
+        if (evt.button === MouseButton.MIDDLE) {
+          restoreToPrevious();
+          dispose();
+        }
+      });
+
+      return false;
+    });
   }
 }