+++ packages/framework/block-std/src/gfx/tool/tool.ts
@@ -31,6 +31,10 @@ export abstract class BaseTool<
     return this.gfx.tool.currentTool$.peek() === this;
   }
 
+  get allowDragWithRightButton() {
+    return false;
+  }
+
   get controller() {
     return this.gfx.tool;
   }