+++ packages/affine/model/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @blocksuite/affine-model
 
+## 0.17.26
+
+### Patch Changes
+
+- cfa436e: ## Feat
+
+  - feat: add pdf viewer to playground (#8650)
+  - feat: add view toggle menu to toolbar of attachment block (#8660)
+  - feat: attachment embedded view supports configurable (#8658)
+
+  ## Fix
+
+  - fix: tidy up after multiple images uploaded (#8671)
+  - fix: unable to add caption for linked document of embedded view (#8670)
+  - fix(blocks): code block should update highlight when theme changed (#8669)
+  - fix(edgeless): allow right click of tool controller (#8652)
+  - fix(database): short name of tag type (#8665)
+  - fix: the content of formula should contain spaces (#8647)
+  - fix: import notion database with title (#8661)
+  - fix(page): bookmark block selected style (#8656)
+  - fix(edgeless): switching shape style and color, the preview does not change in time (#8655)
+  - fix(blocks): size of icons are not consistent in keyboard toolbar (#8653)
+  - fix: use host instead of offsetParent (#8651)
+
+  ## Refactor
+
+  - refactor(database): filter ui (#8611)
+
+  ## Ci
+
+  - ci: renovate pr title lint (#8666)
+
+- Updated dependencies [cfa436e]
+  - @blocksuite/block-std@0.17.26
+  - @blocksuite/global@0.17.26
+  - @blocksuite/inline@0.17.26
+  - @blocksuite/store@0.17.26
+
 ## 0.17.25
 
 ### Patch Changes